# api/openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "Error": {
        "description": "The stable error shape, every non-2xx response, every operation. Mirrors internal/clifmt.CliError's JSON encoding exactly: `code` follows the same exit-code-style bucket (1 = domain/user error such as not-found or conflict, 2 = environment error such as a database outage), and the HTTP status line carries the specific condition.",
        "properties": {
          "code": {
            "enum": [
              1,
              2
            ],
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "remediation"
        ],
        "type": "object"
      },
...
      "PendingDecisionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PendingDecisionRun"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque keyset cursor for the next page; absent on the last page.",
            "type": "string"
          },
          "record_count": {
            "description": "Undecided records across every listed run — the number a stage exit gate compares against zero.",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "record_count"
        ],
        "type": "object"
      },
      "PendingDecisionRecord": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "data": {
            "description": "The record's payload, verbatim — what the decider actually reads.",
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "node_run_id": {
            "type": "string"
          },
          "origin_actor_id": {
            "type": "string"
          },
          "origin_kind": {
            "type": "string"
          },
          "record_type": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "record_type",
          "origin_kind",
          "created_at",
          "data"
        ],
        "type": "object"
      },
      "PendingDecisionRun": {
        "properties": {
          "ledger_version": {
            "description": "The run's current ledger version, read in this same request — what POST /v1alpha1/runs/{id}/reviews wants as `ledger_version`.",
            "format": "int64",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/PendingDecisionRecord"
            },
            "type": "array"
          },
          "run_id": {
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ledger_version",
          "records"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "actorRegistrationBearerAuth": {
        "description": "The shared secret configured as NODES_ACTOR_REGISTRATION_TOKEN_SECRET. Used only by POST /v1alpha1/actors (task t13) — deliberately its own secret, not the decision secret, so an operator can grant registration standing without also granting the power to decide human tasks.",
        "scheme": "bearer",
        "type": "http"
      },
      "adhocRunBearerAuth": {
        "description": "The shared secret configured as NODES_ADHOC_RUN_TOKEN_SECRET. Used only by POST /v1alpha1/adhoc-runs (task t19, gated by the t15 auth-hardening pass, spec c27) — its own secret, so the power to start ad-hoc (often billable) work is granted independently of decision, registration, and event standing.",
        "scheme": "bearer",
        "type": "http"
      },
      "attemptCallbackBearerAuth": {
        "description": "The short-lived callback token minted for exactly one attempt. The route verifies it against the attempt id in the path before loading durable invocation authority or reading the request body.",
        "scheme": "bearer",
        "type": "http"
      },
      "decisionBearerAuth": {
        "description": "The shared secret configured as NODES_HUMAN_DECISION_TOKEN_SECRET. Used by the three operations that record a human decision — POST /v1alpha1/human-tasks/{id}/decision and the two review operations (POST /v1alpha1/runs/{id}/reviews, POST /v1alpha1/reviews/{id}/commit) — because all three write human-authority records into the ledger on whoever presents the token. Every other operation in this API is authless by phase-1 design (spec decision c45).",
        "scheme": "bearer",
        "type": "http"
      },
      "eventBearerAuth": {
        "description": "The shared secret configured as NODES_EVENT_TOKEN_SECRET. Used only by POST /v1alpha1/events (task t10) — its own secret again, so an external system granted the standing to emit signal events (which can resume any signal-parked run) holds neither decision nor registration power.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundActorKeyAuth": {
        "description": "Names the party whose credential is being presented — an actor key (`namespace/name`), which is also the mailbox partition the poll claims from and the completion is scoped to. Required together with inboundDialBearerAuth: an absent or empty header is a 401 before the credential is examined at all. The transport routes hardcode the party KIND to `actor`, so a `host`-kind credential cannot dial in even though issuance will mint one.",
        "in": "header",
        "name": "X-Culture-Nodes-Actor-Key",
        "type": "apiKey"
      },
      "inboundDialBearerAuth": {
        "description": "The credential the control plane ISSUED for one dialling party — the `cnd_`-prefixed value returned exactly once by POST /v1alpha1/inbound/credentials. Presented by a bridge on every poll and every completion, alongside inboundActorKeyAuth; neither header is sufficient alone. Verified against a SHA-256 verifier read from PostgreSQL per call, never cached, which is what makes revocation effective on the next dial.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundIssuanceBearerAuth": {
        "description": "The shared secret configured as NODES_INBOUND_ISSUANCE_TOKEN_SECRET. Used only by POST /v1alpha1/inbound/credentials and POST /v1alpha1/inbound/credentials/revoke (issue #111's dial-in half) — its own secret again, so the standing to mint or end a bridge's IDENTITY is granted independently of decision, registration, event and ad-hoc power. Absent configuration leaves both routes refusing with 401 rather than mounted-but-authless.",
        "scheme": "bearer",
        "type": "http"
...
  },
  "info": {
    "description": "The Culture Nodes control-plane API (docs/initial-design/culture-nodes-prd-spec.md §12.1, §15.1): workflow publication, run orchestration, the append-only work ledger, and review transactions. group: nodes.culture.dev/v1alpha1.\nPhase 1 runs authless by design (spec decision c45): this API carries no authentication or authorization of its own and is meant to run behind a private network. Every operation is reachable by anyone who can reach the listener. Do not expose this API on a public network without a fronting proxy that adds authentication.\nEvery error response uses the Error schema below — the same {code, message, remediation} shape internal/clifmt.CliError renders for the CLI — with the HTTP status line carrying the specific condition. The two reverse-transport routes are the documented exception: POST /v1alpha1/inbound/poll and POST /v1alpha1/inbound/{id}/complete are not mounted through the wrapper that renders Error, so their 400/401 refusals are `text/plain` while their 500s still carry Error. Each of those responses says so where it is declared.",
    "license": {
      "name": "Apache-2.0"
    },
    "summary": "Culture Nodes control-plane API",
    "title": "nodes.culture.dev/v1alpha1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1alpha1/pending-decisions": {
      "get": {
        "description": "Answers \"what is awaiting my decision\", per run and across runs, without a hand-maintained manifest of run ids.\nA record is pending when it carries `proposed` authority and no `review` record names it as its subject — whatever the verdict would have been, since a rejection answers the question as completely as a confirmation does. It is NOT simply `authority = proposed`: records are immutable, so committing a confirmation appends a review record rather than changing the claim, and a decided claim reads `proposed` forever. A record a later record supersedes is excluded too — the correction that replaced it is what should be decided now.\nResults are grouped by run and each group carries that run's current `ledger_version`, which is exactly what POST /v1alpha1/runs/{id}/reviews needs to open a review over them.\nResults are keyset-paginated by `created_at` and `id`. Pass a response's `next_cursor` back as `cursor`; its absence means the final page.\nAn unrecognised query parameter is refused with 400 rather than ignored: a filter silently dropped answers a different question than the one asked while still looking authoritative.",
        "operationId": "listPendingDecisions",
        "parameters": [
          {
            "description": "Only records belonging to this run.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records of this ledger record type (e.g. `claim`).",
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records whose origin actor is this actor.",
            "in": "query",
            "name": "actor_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of records to return (default 200, max 1000).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque keyset cursor from a previous response's `next_cursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingDecisionList"
                }
              }
            },
            "description": "The pending records, grouped by run."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unrecognised query parameter or invalid cursor was supplied."
          }
        },
        "summary": "Every proposed ledger record still awaiting a human decision.",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The API is served relative to whatever host and port `nodes serve` binds.",
      "url": "/"
    }
  ],
...
}